+++ src/test/java/coffee/synyx/auth/authentication/web/LoginControllerDiscoveryDisabledIT.java
@@ -1,0 +1,62 @@
+package coffee.synyx.auth.authentication.web;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import org.junit.runner.RunWith;
+
+import org.springframework.beans.factory.annotation.Autowired;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.ResultActions;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+
+import org.springframework.web.context.WebApplicationContext;
+
+import javax.servlet.Filter;
+
+import static org.hamcrest.core.IsNull.nullValue;
+
+import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
+
+
+/**
+ * @author  Tobias Schneider - schneider@synyx.de
+ */
+@RunWith(SpringRunner.class)
+@SpringBootTest(properties = "coffeenet.discovery.enabled=false")
+public class LoginControllerDiscoveryDisabledIT {
+
+    @Autowired
+    private WebApplicationContext webContext;
+
+    @Autowired
+    private Filter springSecurityFilterChain;
+
+    private MockMvc mockMvc;
+
+    @Before
+    public void setupMockMvc() {
+
+        mockMvc = MockMvcBuilders.webAppContextSetup(webContext).apply(springSecurity(springSecurityFilterChain))
+            .build();
+    }
+
+
+    @Test
+    public void redirectsToDefaultRedirectUriIfLoggedInAndNoInstanceFound() throws Exception {
+
+        ResultActions resultActions = mockMvc.perform(get("/login"));
+        resultActions.andExpect(status().isOk());
+        resultActions.andExpect(view().name("auth/login"));
+        resultActions.andExpect(model().attribute("passwordRecoveryUrl", nullValue()));
+    }
+}